extract Wiring and CurrentDirection into their own files
