+++ src/rotor/current-direction.enum.ts
@@ -1,0 +1,4 @@
+export enum CurrentDirection {
+    FORWARD = "forward",
+    REVERSE = "reverse"
+}
